+++ src/main/java/racingcar/controller/RacingGameController.java
@@ -1,4 +1,9 @@
-package racingcar;
+package racingcar.controller;
+
+import racingcar.domain.Cars;
+import racingcar.domain.RandomNumberGeneratingStrategy;
+import racingcar.view.InputView;
+import racingcar.view.OutputView;
 
 public class RacingGameController {
