+++ src/main/java/racingcar/domain/Car.java
@@ -1,8 +1,8 @@
-package racingcar;
+package racingcar.domain;
 
-import static racingcar.Rule.MOVE_FORWARD_BOUNDARY;
-import static racingcar.Rule.DRIVE_MAXIMUM_VALUE;
-import static racingcar.Rule.DRIVE_MINIMUM_VALUE;
+import static racingcar.domain.Rule.MOVE_FORWARD_BOUNDARY;
+import static racingcar.domain.Rule.DRIVE_MAXIMUM_VALUE;
+import static racingcar.domain.Rule.DRIVE_MINIMUM_VALUE;
 
 public class Car {
