+++ src/main/java/racingcar/domain/FixedNumberGeneratingStrategy.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public class FixedNumberGeneratingStrategy implements NumberGeneratingStrategy {
