+++ src/main/java/racingcar/domain/NumberGeneratingStrategy.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public interface NumberGeneratingStrategy {
