+++ src/main/java/racingcar/view/OutputView.java
@@ -1,8 +1,9 @@
-package racingcar;
+package racingcar.view;
 
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
+import racingcar.domain.Car;
 
 public class OutputView {
